requirements.txt
boto3==1.28.38
pytesseract==0.3.10
Pillow==10.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
openai>=1.10.0
tiktoken>=0.5.1